src/tools/deal-flipper/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
schedule>=1.2
Flask>=3.0
Flask-Cors>=4.0
